refactor: organiza código de ML por responsabilidade
