+++ src/models/mlp.py
@@ -1,0 +1,19 @@
+import torch
+import torch.nn as nn
+
+
+class TelcoMLP(nn.Module):
+    def __init__(self, input_dim: int, hidden_dim: int, dropout_rate: float = 0.2):
+        super().__init__()
+        self.network = nn.Sequential(
+            nn.Linear(input_dim, hidden_dim),
+            nn.ReLU(),
+            nn.Dropout(dropout_rate),
+            nn.Linear(hidden_dim, hidden_dim // 2),
+            nn.ReLU(),
+            nn.Dropout(dropout_rate),
+            nn.Linear(hidden_dim // 2, 1),
+        )
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        return self.network(x)
